+++ tests/test_meta/test_sqlite.py
@@ -10,7 +10,7 @@ import sqlalchemy
 from sqlalchemy.orm import Query, Session
 
 from toggl_api.meta import RequestMethod
-from toggl_api.meta.cache.base_cache import Comparison, TogglQuery
+from toggl_api.meta.cache.base_cache import Comparison, MissingParentError, TogglQuery
 from toggl_api.meta.cache.sqlite_cache import SqliteCache
 from toggl_api.models import TogglTag, TogglTracker, TogglWorkspace
 from toggl_api.models._decorators import UTCDateTime  # noqa: PLC2701
@@ -101,7 +101,7 @@ def test_add_entries_sqlite_parent(meta_object_sqlite, model_data):
     tracker = model_data["tracker"]
     meta_object_sqlite.cache.add_entries(tracker)
     meta_object_sqlite.cache.parent = None
-    with pytest.raises(ValueError, match="Cannot load cache without parent set!"):
+    with pytest.raises(MissingParentError):
         assert tracker in meta_object_sqlite.cache.load_cache()
 
 
@@ -135,7 +135,7 @@ def test_find_sqlite(meta_object_sqlite, model_data):
 @pytest.mark.unit
 def test_find_sqlite_parent(meta_object_sqlite):
     meta_object_sqlite.cache.parent = None
-    with pytest.raises(ValueError, match="Cannot load cache without parent set!"):
+    with pytest.raises(MissingParentError):
         meta_object_sqlite.cache.find_entry({"id": 5})
 
 
@@ -187,7 +187,7 @@ def test_query_sqlite(tracker_object_sqlite, model_data, faker):
     tracker_object_sqlite.cache.commit()
     assert tracker_object_sqlite.load_cache().count() == 11  # noqa: PLR2004
     assert tracker_object_sqlite.query(TogglQuery("name", tracker.name))[0] == tracker
-    assert tracker_object_sqlite.query(TogglQuery("name", list(picked_names))).count() == total_picks
+    assert tracker_object_sqlite.cache.query(TogglQuery("name", list(picked_names))).count() == total_picks
 
 
 @pytest.mark.unit
@@ -204,13 +204,13 @@ def test_query_sqlite_distinct(tracker_object_sqlite, model_data, faker):
         d["timestamp"] = datetime.now(timezone.utc)
         tracker_object_sqlite.save_cache(tracker.from_kwargs(**d), RequestMethod.GET)
 
-    assert tracker_object_sqlite.query(TogglQuery("name", name), distinct=True).count() == 1
+    assert tracker_object_sqlite.cache.query(TogglQuery("name", name), distinct=True).count() == 1
 
 
 @pytest.mark.unit
 def test_query_sqlite_parent(meta_object_sqlite):
     meta_object_sqlite.cache.parent = None
-    with pytest.raises(ValueError, match="Cannot load cache without parent set!"):
+    with pytest.raises(MissingParentError):
         meta_object_sqlite.cache.find_entry({"id": 5})