+++ toggl_api/__init__.py
@@ -1,7 +1,9 @@
+from ._exceptions import DateTimeError, NamingError
 from .client import ClientBody, ClientEndpoint
 from .config import generate_authentication
 from .meta import JSONCache
 from .meta.cache import Comparison, TogglQuery
+from .meta.cache.base_cache import MissingParentError
 from .models import TogglClient, TogglOrganization, TogglProject, TogglTag, TogglTracker, TogglWorkspace
 from .organization import OrganizationEndpoint
 from .project import ProjectBody, ProjectEndpoint
@@ -17,6 +19,8 @@ __typed__ = True
 
 
 __all__ = (  # noqa: RUF022
+    "DateTimeError",
+    "NamingError",
     "ClientBody",
     "ClientEndpoint",
     "Comparison",
@@ -39,4 +43,5 @@ __all__ = (  # noqa: RUF022
     "generate_authentication",
     "TogglOrganization",
     "OrganizationEndpoint",
+    "MissingParentError",
 )