+++ toggl_api/tracker.py
@@ -1,12 +1,14 @@
 from __future__ import annotations
 
 import logging
+import warnings
 from dataclasses import dataclass, field
 from datetime import date, datetime, timedelta, timezone
 from typing import Any, Final, Literal, Optional
 
 from httpx import HTTPStatusError, codes
 
+from toggl_api._exceptions import NamingError
 from toggl_api.meta import BaseBody, RequestMethod, TogglCachedEndpoint
 from toggl_api.models import TogglTracker
 from toggl_api.utility import format_iso
@@ -20,7 +22,7 @@ class TrackerBody(BaseBody):
 
     Examples:
         >>> TrackerBody(description="What a wonderful tracker description!", project_id=2123132)
-        TrackerBody(description='What a wonderful tracker description!', project_id=2123132)
+        TrackerBody(description="What a wonderful tracker description!", project_id=2123132)
     """
 
     description: Optional[str] = field(default=None)
@@ -88,7 +90,7 @@ class TrackerBody(BaseBody):
         return body
 
 
-class TrackerEndpoint(TogglCachedEndpoint):
+class TrackerEndpoint(TogglCachedEndpoint[TogglTracker]):
     """Endpoint for modifying and creating trackers.
 
     See the [UserEndpoint][toggl_api.UserEndpoint] for _GET_ specific requests.
@@ -158,6 +160,9 @@ class TrackerEndpoint(TogglCachedEndpoint):
 
         Args:
             tracker: Tracker object with ID to delete.
+
+        Raises:
+            HTTPStatusError: If anything thats not a '404' or 'ok' code is returned.
 
         Returns:
             None: If the tracker was deleted or not found at all.
@@ -235,14 +240,22 @@ class TrackerEndpoint(TogglCachedEndpoint):
                 to -1 for a running tracker.
 
         Raises:
-            ValueError: Description must be set in order to create a new tracker.
+            NamingError: Description must be set in order to create a new tracker.
 
         Returns:
             TogglTracker | None: The tracker that was created.
         """
-        if not isinstance(body.description, str) or not body.description:
+        if not isinstance(body.description, str):
+            warnings.warn(
+                "DEPRECATED: 'TypeError' is being swapped for a 'NamingError'.",
+                DeprecationWarning,
+                stacklevel=2,
+            )
             msg = "Description must be set in order to create a tracker!"
             raise TypeError(msg)
+        if not body.description:
+            msg = "Description must be set in order to create a tracker!"
+            raise NamingError(msg)
 
         if body.start is None and body.start_date is None:
             body.start = datetime.now(tz=timezone.utc)